tools/cobalt-strike-extractor/pyproject.toml
[tool.poetry]
name = "elastic-cobalt-strike-extractor"
version = "0.1.0"
description = "Python script to collect Cobalt Strike memory data generated by security events from an Elasticsearch cluster, extracts the configuration from the CS beacon, and writes the data back to Elasticsearch."
authors = ["Dan Stepanic <daniel.stepanic@elastic.co>", "Derek Ditch <dcode@elastic.co>"]
readme = "README.md"
packages = [{ include = "elastic" }]

[tool.poetry.scripts]
cobalt-strike-extractor = 'elastic.cobalt_strike_extractor.console:main'
elastic-extractor = 'elastic.cobalt_strike_extractor.util.elastic_extractor:main'

[tool.poetry.dependencies]
python = "^3.9"
elasticsearch = "^7.16.2"
pefile = "^2021.9.3"
libcsce = "^0.1.0"
ecs-logging = "^1.1.0"
scalpl = "^0.4.2"
docopt = "^0.6.2"
"ruamel.yaml" = "^0.17.20"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0"
black = "^21.10b0"
flake8 = "<4"
pytest-black = "^0.3.12"
pytest-flake8 = "^1.0.7"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest-isort = "^2.0.0"
pytest-datadir = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--black --isort --flake8"
testpaths = ["."]
flake8-max-line-length = 89
flake8-ignore = "E203 E266 E501 W503"

[tool.pyright]
include = ["elastic", "tests"]
exclude = ["**/__pycache__"]
